+++ src/Bootstrap.php
@@ -6,40 +6,36 @@ namespace Hirtz\Anakin;
 
 use Hirtz\Anakin\Assets\AnakinAssetBundle;
 use Hirtz\Skeleton\Assets\AdminAssetBundle;
-use Hirtz\Skeleton\Assets\TinyMceAssetBundle;
+use Hirtz\Skeleton\Models\User;
 use Hirtz\Skeleton\Modules\Admin\Module;
 use Hirtz\Skeleton\Web\Application;
 use Yii;
+use yii\base\ActionEvent;
 use yii\base\BootstrapInterface;
 use yii\base\Event;
 use yii\i18n\PhpMessageSource;
-use yii\web\View;
 
 class Bootstrap implements BootstrapInterface
 {
     /**
-     * @param Application<\Hirtz\Skeleton\Models\User> $app
+     * @param Application<User> $app
      */
     public function bootstrap($app): void
     {
         Yii::setAlias('@anakin', __DIR__);
-        Yii::setAlias('@skeleton/mail/layouts/html', '@anakin/views/layouts/mail');
+        Yii::setAlias('@skeleton/../resources/mail/layouts/html', '@anakin/../resources/mail/layouts/html');
 
         $app->getI18n()->translations['anakin'] ??= [
             'class' => PhpMessageSource::class,
             'basePath' => '@anakin/messages',
         ];
 
-        Yii::$app->params['email'] ??= 'hello@anakin.co';
+        $app->params['email'] ??= 'hello@anakin.co';
 
-        Yii::$app->extendComponents([
+        $app->extendComponents([
             'assetManager' => [
                 'bundles' => [
-                    TinyMceAssetBundle::class => [
-//                        'sourcePath' => '@anakin/../resources/assets/tinymce/skins/',
-                    ],
                     AdminAssetBundle::class => [
-//                        'css' => [], // Remove admin CSS, the Anakin theme will register its own CSS
                         'faviconOptions' => [
                             'href' => '/images/favicons/favicon-32x32.png',
                             'sizes' => '32x32',
@@ -54,19 +50,17 @@ class Bootstrap implements BootstrapInterface
                         '@skeleton/../resources/views/admin/views/dashboard' => '@anakin/../resources/views/dashboard',
                     ],
                 ],
-            ]
+            ],
         ]);
 
-        // Todo
-        //        Event::on(TinyMceEditor::class, TinyMceEditor::EVENT_INIT, function ($event) {
-        //            $asset = AnakinAsset::register(Yii::$app->getView());
-        //            $event->sender->clientOptions['content_css'] ??= "$asset->baseUrl/css/tinymce.min.css";
-        //        });
+        Event::on(Module::class, Module::EVENT_BEFORE_ACTION, function (ActionEvent $event): void {
+//            /** @var Module $module */
+//            $module = $event->sender;
+            $view = $event->action->controller->getView();
 
-        Event::on(View::class, View::EVENT_BEGIN_PAGE, function (): void {
-//            if (Yii::$app->controller?->module instanceof Module || Yii::$app->controller?->module->module instanceof Module) {
-//                AnakinAssetBundle::register(Yii::$app->getView());
-//            }
+            $view->on($view::EVENT_BEGIN_PAGE, function () use ($module, $view) {
+                AnakinAssetBundle::register($view);
+            });
         });
     }
 }